refactor: componentize year input
