+++ src/components/SpreadsheetUpload.tsx
@@ -14,11 +14,7 @@
 
 import type React from "react";
 import FileUpload from "@/components/FileUpload";
-import { Button } from "./ui/button";
-import { Input } from "./ui/input";
-import { ChevronLeft } from "lucide-react";
-import { ChevronRight } from "lucide-react";
-import { useEffect, useState } from "react";
+import YearInput from "@/components/YearInput";
 
 type UploadProps = {
     file?: File;
@@ -33,38 +29,6 @@ export default function SpreadsheetUpload({
     year,
     setYear,
 }: UploadProps) {
-    const [yearStr, setYearStr] = useState("");
-    const currYear = Number(yearStr);
-
-    const handleYearInput = (e: React.ChangeEvent<HTMLInputElement>) => {
-        if (year) {
-            const value = e.target.value;
-            const newYear = Number(value);
-            setYear(newYear);
-            setYearStr(value);
-        }
-    };
-
-    const incrementYear = () => {
-        if (year && currYear < 2100) {
-            const newYear = currYear + 1;
-            setYear(newYear);
-            setYearStr(String(newYear));
-        }
-    };
-
-    const decrementYear = () => {
-        if (year && currYear > 2000) {
-            const newYear = currYear - 1;
-            setYear(newYear);
-            setYearStr(String(newYear));
-        }
-    };
-
-    useEffect(() => {
-        setYearStr(String(year));
-    }, [year]);
-
     return (
         <div>
             <div className="flex flex-col gap-4">
@@ -76,35 +40,8 @@ export default function SpreadsheetUpload({
 
                 <h2 className="text-base mt-4">Year</h2>
 
-                <div className="flex items-center w-[180px] rounded-md overflow-hidden shadow-sm">
-                    {/* Left Arrow Button */}
-                    <Button
-                        variant="outline"
-                        onClick={decrementYear}
-                        className="h-9 w-10 flex items-center justify-center rounded-none border-r-0 shadow-none"
-                    >
-                        <ChevronLeft className="h-4 w-4" />
-                    </Button>
+                <YearInput year={year} setYear={setYear} />
 
-                    {/* Year Input */}
-                    <Input
-                        type="number"
-                        id="year"
-                        name="Year"
-                        value={yearStr}
-                        onChange={handleYearInput}
-                        className="h-9 w-[100px] text-center rounded-none border-y border-x-0 shadow-none focus-visible:ring-0 [appearance:textfield] [&::-webkit-outer-spin-button]:appearance-none [&::-webkit-inner-spin-button]:appearance-none"
-                    />
-
-                    {/* Right Arrow Button */}
-                    <Button
-                        variant="outline"
-                        onClick={incrementYear}
-                        className="h-9 w-10 flex items-center justify-center rounded-none border-l-0 shadow-none"
-                    >
-                        <ChevronRight className="h-4 w-4" />
-                    </Button>
-                </div>
                 <div className="m-5" />
                 <FileUpload fileInfo={file} setFileInfo={setFile} />
             </div>
